move rand to utils
